requirements.txt
Django==1.10.3
django-model-utils==2.6
djangorestframework==3.5.3
psycopg2==2.6.2
